Embed sub-program source as a resource
